requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.109.0
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
h11==0.14.0
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
mobile-sam @ git+https://github.com/dhkim2810/MobileSAM.git@a00e905633954fa29c98c1d93f052847bc5fa0e2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
sympy==1.12
timm==0.9.12
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0
